[CHORE] Move assert-all-deprecations to test-infra and apply to encapsulation apps
