+++ packages/unpublished-test-infra/addon-test-support/assert-all-deprecations.js
@@ -1,0 +1,39 @@
+import { DEBUG } from '@glimmer/env';
+import QUnit from 'qunit';
+
+export default function configureAssertAllDeprecations() {
+  QUnit.begin(() => {
+    function assertAllDeprecations(assert) {
+      if (typeof assert.test.expected === 'number') {
+        assert.test.expected += 1;
+      }
+      assert.expectNoDeprecation(undefined, undefined, deprecation => {
+        // only assert EmberData deprecations
+        const id = deprecation.options.id.toLowerCase();
+        const isEmberDataDeprecation =
+          id.includes('ds.') ||
+          id.includes('emberdata') ||
+          id.includes('ember-data') ||
+          id.includes('mismatched-inverse-relationship-data-from-payload');
+
+        if (!isEmberDataDeprecation) {
+          // eslint-disable-next-line no-console
+          console.warn('Detected Non-Ember-Data Deprecation:', deprecation.message, deprecation.options.stacktrace);
+        }
+
+        return isEmberDataDeprecation;
+      });
+    }
+    // ensure we don't regress quietly
+    // this plays nicely with `expectDeprecation`
+    if (DEBUG) {
+      QUnit.config.modules.forEach(mod => {
+        const hooks = (mod.hooks.afterEach = mod.hooks.afterEach || []);
+
+        if (mod.tests.length !== 0) {
+          hooks.unshift(assertAllDeprecations);
+        }
+      });
+    }
+  });
+}
