+++ packages/unpublished-test-infra/addon-test-support/legacy.js
@@ -1,0 +1,27 @@
+import { wait, asyncEqual, invokeAsync } from './async';
+import QUnit from 'qunit';
+
+const { assert } = QUnit;
+
+export default function additionalLegacyAsserts() {
+  assert.wait = wait;
+  assert.asyncEqual = asyncEqual;
+  assert.invokeAsync = invokeAsync;
+
+  assert.assertClean = function(promise) {
+    return promise.then(
+      this.wait(record => {
+        this.equal(record.get('hasDirtyAttributes'), false, 'The record is now clean');
+        return record;
+      })
+    );
+  };
+
+  assert.contains = function(array, item) {
+    this.ok(array.indexOf(item) !== -1, `array contains ${item}`);
+  };
+
+  assert.without = function(array, item) {
+    this.ok(array.indexOf(item) === -1, `array doesn't contain ${item}`);
+  };
+}
